+++ src/pages/home/HomePage.tsx
@@ -1,7 +1,7 @@
 import { type FC } from 'react';
 import { CardList, SearchBar } from '~compos/_index';
 
-export const Home: FC = () => {
+export const HomePage: FC = () => {
   return (
     <div data-testid="home">
       <SearchBar />
